+++ packages/frameworks-solid-start/src/client.ts
@@ -42,6 +42,7 @@ export async function signIn<
   const _signInUrl = `${signInUrl}?${new URLSearchParams(authorizationParams)}`;
 
   // TODO: Handle custom base path
+  // TODO: Remove this since Sveltekit offers the CSRF protection via origin check
   const csrfTokenResponse = await fetch("/api/auth/csrf");
   const { csrfToken } = await csrfTokenResponse.json();
