--- packages/frameworks-solid-start/tsconfig.m.json
@@ -1,15 +1,0 @@
-{
-  "compilerOptions": {
-    "declaration": true,
-    "allowSyntheticDefaultImports": true,
-    "target": "esnext",
-    "newLine": "LF",
-    "moduleResolution": "Node",
-    "strict": false,
-    "jsx": "preserve",
-    "jsxImportSource": "solid-js",
-    "module": "esnext",
-    "types": ["vite/client", "node"]
-  },
-  "exclude": ["node_modules", "dist"]
-}
